Extracted from Main, Created HouseVotes, and Added Loss Functions
I removed unnecessary files, I created the HouseVotesData and HouseVotesDataNoise. TrainClassLoss is where all training, classifying, and loss functions take place. The Data Main does nothing but take the data, put it in classes, and then send it to TrainClassLoss. The loss function used is the confusion matrix for each.